changed conversion to use huggingface
